requirements.txt
torch==1.13.0
Pillow==9.2.0
torchvision==0.14.0
matplotlib==3.6.2
scikit-learn==1.2.1
pandas==1.5.3